# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /engine:
    get:
      summary: "find engines"
      tags:
        - "engine"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the engines matching the filters or all engines if no filters are specified."
          schema:
            type: object
            properties:
              results:
                type: "array"
                items:
                  $ref: "#/definitions/Engine"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
    post:
      summary: "create an engine"
      tags:
        - "engine"
      parameters:
        - name: "engine"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Engine"
      responses:
        "200":
          description: "the engine"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Engine"
  /engine/{engine_id}:
    parameters:
      - name: "engine_id"
        type: "string"
        in: "path"
        required: true
    get:
      summary: "get an engine instance"
      tags:
        - "engine"
      responses:
        "200":
          description: "the engine"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Engine"
    put:
      summary: "update an engine"
      tags:
        - "engine"
      parameters:
        - name: "engine"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Engine"
      responses:
        "200":
          description: "the engine"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Engine"
    delete:
      summary: "delete an engine"
      tags:
        - "engine"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
  /engine/{engine_id}/subgraph_definition:
    parameters:
      - name: "engine_id"
        type: "string"
        in: "path"
        required: true
    get:
      summary: "return the subgraph definition for the engine"
      tags:
        - "engine"
        - "subgraph definition"
      responses:
        "200":
          description: "the subgraph definition"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/SubGraphDefinition"
    post:
      summary: "create the subgraph definition for the engine"
      tags:
        - "engine"
        - "subgraph definition"
      parameters:
        - name: "subgraph_definition"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/SubGraphDefinition"
      responses:
        "200":
          description: "the subgraph definition"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/SubGraphDefinition"
  /engine/action/{action_id}/checkout:
    put:
      summary: "Checkout the action"
      tags:
        - "engine"
        - "action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/ActionContext"
        "400":
          description: "engine action is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResult"
        "404":
          description: "engine or datastore not found"
          schema:
            $ref: "#/definitions/ErrorResult"
  /engine/action/{action_id}/checkin:
    put:
      summary: "Checkin the action"
      tags:
        - "engine"
        - "action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
        - name: "action_result"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/ActionResult"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
        "400":
          description: "engine action is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResult"
        "404":
          description: "engine or datastore not found"
          schema:
            $ref: "#/definitions/ErrorResult"
  /event:
    post:
      summary: "create an event"
      tags:
        - "event"
      parameters:
        - name: "event"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Event"
      responses:
        "200":
          description: returns the event
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Event"
    get:
      summary: "find events"
      tags:
        - "event"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the events matching the filters or all events if no filters are specified."
          schema:
            type: object
            properties:
              results:
                type: "array"
                items:
                  $ref: "#/definitions/Event"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
  /event/{event_id}:
    parameters:
      - name: "event_id"
        type: "string"
        in: "path"
        required: true
    get:
      summary: "get an event"
      tags:
        - "event"
      responses:
        "200":
          description: returns the event
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Event"
    put:
      summary: "update an event"
      tags:
        - "event"
      parameters:
        - name: "event"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Event"
      responses:
        "200":
          description: returns the event
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Event"
    patch:
      summary: "patch update an event"
      tags:
        - "event"
      parameters:
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: returns the event
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Event"
    delete:
      summary: "delete an event"
      tags:
        - "event"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
...
  /subgraph_definition/{subgraph_definition_id}:
    parameters:
      - name: "subgraph_definition_id"
        type: "string"
        in: "path"
        required: true
    get:
      summary: "get the subgraph definition"
      tags:
        - "subgraph definition"
      responses:
        "200":
          description: "the subgraph definition"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/SubGraphDefinition"
    delete:
      summary: "delete the subgraph definition"
      tags:
        - "subgraph definition"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
...
parameters:
  limitParam:
    name: "limit"
    type: "integer"
    in: "query"
    description: |
      Specify the maximum number of records that should be returned by this request.
  offsetParam:
    name: "offset"
    type: "integer"
    in: "query"
    description: |
      Specify the starting index for the next chunk of results to be returned by this request.
  filtersParam:
    name: filters
    in: query
    type: string
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of Filter objects (as defined in
      the definitions section).
...
  patchParam:
    name: "patch"
    description: "a JSON Patch object"
    in: "body"
    required: true
    schema:
      $ref: '#/definitions/JSONPatch'
#
# Definitions are auto-generated by dart.tool.export_swagger_definitions
# Do not manually edit!
#
definitions:
...
  ActionContext:
    type: object
...
  ActionResult:
    type: object
  ActionType:
    properties:
      description:
        type: string
      name:
        type: string
      params_json_schema:
        type: object
    type: object
...
  Engine:
    additionalProperties: false
    properties:
      created:
        type: string
      data:
        $ref: '#/definitions/EngineData'
      id:
        type: string
      updated:
        type: string
      version_id:
        type: integer
    required:
    - data
    type: object
  EngineData:
    additionalProperties: false
    properties:
      description:
        type: string
      ecs_task_definition:
        type: object
      ecs_task_definition_arn:
        type: string
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      options_json_schema:
        type: object
      supported_action_types:
        items:
          $ref: '#/definitions/ActionType'
        minItems: 1
        type: array
        x-schema-form:
          title: '{{ value.name }}'
          type: tabarray
      tags:
        default: []
        items:
          maxLength: 30
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
    required:
    - name
    type: object
  ErrorResult:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
  Event:
    additionalProperties: false
    properties:
      created:
        type: string
      data:
        $ref: '#/definitions/EventData'
      id:
        type: string
      updated:
        type: string
      version_id:
        type: integer
    required:
    - data
    type: object
  EventData:
    additionalProperties: false
    properties:
      description:
        type: string
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      state:
        default: INACTIVE
        enum:
        - ACTIVE
        - INACTIVE
        type: string
      tags:
        default: []
        items:
          maxLength: 30
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
    required:
    - name
    type: object
...
  JSONPatch:
    description: JSON Patch object as defined by http://json.schemastore.org/json-patch
    type: object
...
  OKResult:
    properties:
      results:
        enum:
        - OK
        type: string
    type: object
...
  SubGraphDefinition:
    type: object
...